Reimplement basic GLFW windowing
