+++ flutter-glfw/Cargo.toml
@@ -22,3 +22,4 @@ lazy_static = "1.3.0"
 priority-queue = "^0.6"
 locale_config = "^0.2"
 unic-locale = "^0.4"
+parking_lot = "^0.10.0"
